requirements.txt
torch>=1.9.0
torchvision>=0.10.0
trimesh>=3.9.0
pytorch3d>=0.6.0
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.4.0
opencv-python>=4.5.0
pyyaml>=5.4.0
tqdm>=4.60.0
tensorboard>=2.5.0
plotly>=5.0.0
requests>=2.25.0
